Update
Added variable numUnfollows. Exchange 000 for number of follows you need per run.
Added len(id) if statment. Only unfollows if user is following more than trigger point. I set trigger point to 4000 follows. Max follows for unverified is 5000.

+++ unfollow50.py
@@ -16,6 +16,7 @@ if __name__ == "__main__":
     api = initialize_api()
 
     me = "XXX"
+    numUnfollows = 000
 
     logger.info("Gathering Friend's IDs")
     ids = api.friends_ids(me)
@@ -23,12 +24,14 @@ if __name__ == "__main__":
     logger.info("Reversing ID Array")
     ids.reverse()
 
-    logger.info("Starting Unfollow Loop")
-    i=0
-    while i<49:
-        api.destroy_friendship(ids[i])
-        logger.info("Sleeping for 30 seconds")
-        time.sleep(30)
-        i += 1
+    if (len(ids) >= 4000):
+        logger.info("Starting Unfollow")
+        i=0
+        while i<numUnfollows:
+            api.destroy_friendship(ids[i])
+            logger.info("Sleeping for 30 seconds")
+            time.sleep(30)
+            i += 1
+        logger.info("Completed removing numUnfollows")
 
     logger.info("Finished")
